+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Added
 
-- `moonstone/ui/IconButton` component, transferred from `ui/IconButton`
+- `UiIconButton` component, transferred from `ui/IconButton`
 - `moonstone/ui/RadioDecorator` component, transferred from `ui/RadioDecorator`
 - `moonstone/ui/Scroller.UiScrollerBase` component, transferred from `ui/Scroller.UiScrollerBase`
 - `moonstone/ui/SlotItem` component, transferred from `ui/SlotItem`